requirements.txt
crewai>=0.16.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
google-generativeai>=0.3.2
Pillow>=10.0.0
python-latex>=1.0.0
pathlib>=1.0.1
logging>=0.5.1.2
langchain>=0.0.305
langchain-community>=0.0.16
langchain-huggingface>=0.0.1
langchain-core>=0.1.5
faiss-cpu>=1.7.4
huggingface_hub>=0.19.0
sentence-transformers>=2.2.2
mistralai>=0.0.8
requests>=2.31.0
reportlab>=4.0.4 